cleaned up code styled addComment component
