+++ instagram-app/src/components/PostContainer/PostContainer.css
@@ -27,5 +27,10 @@
 }
 
 .add-comment {
-    border-top: 1px solid lightgray;
+    display: flex;
+    justify-content: center;
+    width: 50%;
+    text-align: center;
+    margin-left: 20%;
+    border: 1px solid lightgray;
 }
